Create Clothing Store Database and Stats
This is my store database. I've created a table using SQL to categorize the numbered item, the name, quantity, and price of each article of clothing. 
Notice how I've descibed each piece of code. This allows me to see how much I have left of each clothing and which products are running out.